refactor- pull out common component from RegisterPage (AuthInput, AuthInputAlert, AuthInputWordCount)
