+++ src/Pages/RegisterPage/RegisterPage.module.scss
@@ -75,10 +75,6 @@
   margin-top: 4px;
 }
 
-.authinput-alertmsg {
-  color: var(--color-error);
-}
-
 .auth-button {
   margin-top: 24px;
   width: 25.4%;
